refactor(benchmark): split into types, lib/, commands/ with @app/ alias imports
