+++ src/benchmark/commands/add.ts
@@ -1,0 +1,88 @@
+import * as p from "@clack/prompts";
+import { getCustomSuites, BUILTIN_SUITES, saveCustomSuites } from "@app/benchmark/lib/suites";
+import type { AddOptions, BenchmarkCommand, BenchmarkSuite } from "@app/benchmark/types";
+
+export async function cmdAdd(name: string, commandPairs: string[], opts: AddOptions = {}): Promise<void> {
+    if (BUILTIN_SUITES.some((s) => s.name === name)) {
+        p.log.error(`Cannot overwrite built-in suite "${name}".`);
+        process.exit(1);
+    }
+
+    // Parse --prepare-for pairs into a lookup: label → prepare command
+    const prepareForMap = new Map<string, string>();
+
+    for (const pair of opts.prepareFor ?? []) {
+        const eqIdx = pair.indexOf("=");
+
+        if (eqIdx === -1) {
+            p.log.error(`Invalid --prepare-for format: "${pair}". Expected "label=command".`);
+            process.exit(1);
+        }
+
+        prepareForMap.set(pair.slice(0, eqIdx), pair.slice(eqIdx + 1));
+    }
+
+    const commands: BenchmarkCommand[] = [];
+
+    for (const pair of commandPairs) {
+        const colonIdx = pair.indexOf(":");
+
+        if (colonIdx === -1) {
+            p.log.error(`Invalid format: "${pair}". Expected "label:command".`);
+            process.exit(1);
+        }
+
+        const label = pair.slice(0, colonIdx);
+        const cmd: BenchmarkCommand = {
+            label,
+            cmd: pair.slice(colonIdx + 1),
+        };
+
+        const perCmdPrepare = prepareForMap.get(label);
+
+        if (perCmdPrepare) {
+            cmd.prepare = perCmdPrepare;
+        }
+
+        commands.push(cmd);
+    }
+
+    if (commands.length < 2) {
+        p.log.error("A benchmark suite needs at least 2 commands to compare.");
+        process.exit(1);
+    }
+
+    const suite: BenchmarkSuite = { name, commands };
+
+    if (opts.runs) {
+        suite.runs = opts.runs;
+    }
+
+    if (opts.warmup !== undefined) {
+        suite.warmup = opts.warmup;
+    }
+
+    if (opts.setup) {
+        suite.setup = opts.setup;
+    }
+
+    if (opts.prepare) {
+        suite.prepare = opts.prepare;
+    }
+
+    if (opts.cleanup) {
+        suite.cleanup = opts.cleanup;
+    }
+
+    const custom = await getCustomSuites();
+    const existing = custom.findIndex((s) => s.name === name);
+
+    if (existing >= 0) {
+        custom[existing] = suite;
+    } else {
+        custom.push(suite);
+    }
+
+    await saveCustomSuites(custom);
+    p.log.success(`Suite "${name}" saved with ${commands.length} commands.`);
+}
